refactor(users): entities/user を pages/users 内へマージ
steiger の fsd/insignificant-slice 警告（entities/user は pages/users
からしか参照されない）を解消するため、User 型と関連 import を pages/users 配下に集約。
動作変更なし、テスト 345 件 / lint:fsd / lint / typecheck すべて pass。
