--- src/entities/user/index.ts
@@ -1,1 +1,0 @@
-export type { User } from "./model/user";
