--- src/entities/user/model/user.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id: number;
-  uuid: string;
-  first_name: string;
-  last_name: string;
-};
