+++ src/pages/users/model/user.ts
@@ -1,4 +1,9 @@
-import type { User } from "@/entities/user";
+export type User = {
+  id: number;
+  uuid: string;
+  first_name: string;
+  last_name: string;
+};
 
 export type UsersResponse = {
   users: User[];
